pvmfw: Enable booting of the kernel through EFI
Implement EFI stub in pvmfw to add UEFI support and enable booting og
the guest kernel through it.

WIP. This commit is not intended to be merged.

Test: adb shell /apex/com.android.virt/bin/crosvm run
--unmap-guest-memory-on-fork --disable-sandbox
--protected-vm-with-firmware /data/local/tmp/custom_pvmfw
--initrd /apex/com.android.virt/etc/microdroid_initrd_debuggable.img
/data/local/tmp/microdroid_kernel_supports_uefi_boot

+++ guest/pvmfw/src/uefi_deps.rs
@@ -94,7 +94,7 @@ pub struct Event(u32);
 
 #[derive(Copy, Clone, Debug)]
 #[repr(C)]
-pub struct Char16(u32);
+pub struct Char16(u16);
 
 #[derive(Copy, Clone, Debug)]
 #[repr(C)]
